requirements.txt
mysql.connector
subprocess
winreg
psutil